# apps/api/docs/openapi_v1.0.0.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CollegeERPAPI",
    "version": "1.0.0",
    "description": "\nCollege ERP System API\n\n**Generated:** 2026-02-03T12:39:45.480081\n\n**Locked Domains:**\n- Admission (v1.0.0) - Contracts locked with strict validation\n\n**Status:**\n- \u2705 Admission: Ready for frontend integration\n- \u23f3 Finance: In progress\n- \u23f3 Academic: In progress\n- \u23f3 Student: In progress\n\n**Breaking Changes Policy:**\n- Enum additions: Safe (backward compatible)\n- New optional fields: Safe\n- Required field changes: Requires migration + deprecation notice\n- Enum removals: 6-month deprecation period\n- Field type changes: Major version bump\n        "
  },
  "paths": {
...
    "/api/v1/campus/hostels/hostels/blocks": {
      "post": {
        "tags": [
          "campus",
          "Campus - Hostel",
          "Hostel Residency"
        ],
        "summary": "Create Hostel Block",
        "description": "Create a new hostel block",
        "operationId": "create_hostel_block_api_v1_campus_hostels_hostels_blocks_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostelBlockCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostelBlock"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "campus",
          "Campus - Hostel",
          "Hostel Residency"
        ],
        "summary": "List Hostel Blocks",
        "description": "List all hostel blocks",
        "operationId": "list_hostel_blocks_api_v1_campus_hostels_hostels_blocks_get",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/HostelType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HostelBlock"
                  },
                  "title": "Response List Hostel Blocks Api V1 Campus Hostels Hostels Blocks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campus/hostels/hostels/blocks/{block_id}/statistics": {
      "get": {
        "tags": [
          "campus",
          "Campus - Hostel",
          "Hostel Residency"
        ],
        "summary": "Get Block Statistics",
        "description": "Get residency statistics for a block",
        "operationId": "get_block_statistics_api_v1_campus_hostels_hostels_blocks__block_id__statistics_get",
        "parameters": [
          {
            "name": "block_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Block Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campus/hostels/hostels/rooms": {
      "post": {
        "tags": [
          "campus",
          "Campus - Hostel",
          "Hostel Residency"
        ],
        "summary": "Create Room",
        "description": "Create a room in a hostel block",
        "operationId": "create_room_api_v1_campus_hostels_hostels_rooms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostelRoomCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostelRoom"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/campus/hostels/hostels/rooms/available": {
      "get": {
        "tags": [
          "campus",
          "Campus - Hostel",
          "Hostel Residency"
        ],
        "summary": "Get Available Rooms",
        "description": "Get available rooms for allocation",
        "operationId": "get_available_rooms_api_v1_campus_hostels_hostels_rooms_available_get",
        "parameters": [
          {
            "name": "block_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Block Id"
            }
          },
          {
            "name": "room_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RoomType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Room Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HostelRoom"
                  },
                  "title": "Response Get Available Rooms Api V1 Campus Hostels Hostels Rooms Available Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campus/hostels/hostels/allocations": {
      "post": {
        "tags": [
          "campus",
          "Campus - Hostel",
          "Hostel Residency"
        ],
        "summary": "Allocate Room",
        "description": "Allocate a bed to a student",
        "operationId": "allocate_room_api_v1_campus_hostels_hostels_allocations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomAllocationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BedAllocation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/campus/hostels/hostels/allocations/{allocation_id}/vacate": {
      "post": {
        "tags": [
          "campus",
          "Campus - Hostel",
          "Hostel Residency"
        ],
        "summary": "Vacate Room",
        "description": "Deactivate a room allocation (Student vacating)",
        "operationId": "vacate_room_api_v1_campus_hostels_hostels_allocations__allocation_id__vacate_post",
        "parameters": [
          {
            "name": "allocation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Allocation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BedAllocation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campus/hostels/hostels/complaints": {
      "post": {
        "tags": [
          "campus",
          "Campus - Hostel",
          "Hostel Residency"
        ],
        "summary": "Create Complaint",
        "description": "Submit a residency-related complaint/maintenance request",
        "operationId": "create_complaint_api_v1_campus_hostels_hostels_complaints_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplaintCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostelComplaint"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "campus",
          "Campus - Hostel",
          "Hostel Residency"
        ],
        "summary": "List Complaints",
        "description": "List hostel residency complaints",
        "operationId": "list_complaints_api_v1_campus_hostels_hostels_complaints_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ComplaintStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HostelComplaint"
                  },
                  "title": "Response List Complaints Api V1 Campus Hostels Hostels Complaints Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BedAllocation": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "room_id": {
            "type": "integer",
            "title": "Room Id"
          },
          "bed_number": {
            "type": "string",
            "title": "Bed Number"
          },
          "student_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Id"
          },
          "faculty_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Faculty Id"
          },
          "allocation_date": {
            "type": "string",
            "format": "date-time",
            "title": "Allocation Date"
          },
          "deallocation_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deallocation Date"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "room_id",
          "bed_number"
        ],
        "title": "BedAllocation",
        "description": "Bed Assignment - Specific student residency record"
      },
...
      "ComplaintCreate": {
        "properties": {
          "room_id": {
            "type": "integer",
            "title": "Room Id"
          },
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "default": "MEDIUM"
          }
        },
        "type": "object",
        "required": [
          "room_id",
          "student_id",
          "category",
          "description"
        ],
        "title": "ComplaintCreate"
      },
      "ComplaintStatus": {
        "type": "string",
        "enum": [
          "OPEN",
          "IN_PROGRESS",
          "RESOLVED",
          "CLOSED"
        ],
        "title": "ComplaintStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HostelBlock": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/HostelType"
          },
          "total_floors": {
            "type": "integer",
            "title": "Total Floors"
          },
          "warden_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warden Name"
          },
          "contact_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Number"
          }
        },
        "type": "object",
        "required": [
          "name",
          "type",
          "total_floors"
        ],
        "title": "HostelBlock",
        "description": "Hostel Block - Infrastructure Owner for Residency"
      },
      "HostelBlockCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/HostelType"
          },
          "total_floors": {
            "type": "integer",
            "title": "Total Floors"
          },
          "warden_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warden Name"
          },
          "contact_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Number"
          }
        },
        "type": "object",
        "required": [
          "name",
          "type",
          "total_floors"
        ],
        "title": "HostelBlockCreate"
      },
      "HostelComplaint": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "room_id": {
            "type": "integer",
            "title": "Room Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "default": "MEDIUM"
          },
          "status": {
            "$ref": "#/components/schemas/ComplaintStatus",
            "default": "OPEN"
          },
          "resolution_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Note"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "room_id",
          "category",
          "description"
        ],
        "title": "HostelComplaint",
        "description": "Hostel Complaint - Residency-related service request"
      },
      "HostelRoom": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "block_id": {
            "type": "integer",
            "title": "Block Id"
          },
          "room_number": {
            "type": "string",
            "title": "Room Number"
          },
          "floor": {
            "type": "integer",
            "title": "Floor"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "current_occupancy": {
            "type": "integer",
            "title": "Current Occupancy",
            "default": 0
          },
          "room_type": {
            "$ref": "#/components/schemas/RoomType",
            "default": "NON_AC"
          },
          "monthly_rent": {
            "type": "number",
            "title": "Monthly Rent"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "block_id",
          "room_number",
          "floor",
          "capacity",
          "monthly_rent"
        ],
        "title": "HostelRoom",
        "description": "Hostel Room - Resident Living Space"
      },
      "HostelRoomCreate": {
        "properties": {
          "block_id": {
            "type": "integer",
            "title": "Block Id"
          },
          "room_number": {
            "type": "string",
            "title": "Room Number"
          },
          "floor": {
            "type": "integer",
            "title": "Floor"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "room_type": {
            "$ref": "#/components/schemas/RoomType",
            "default": "NON_AC"
          },
          "monthly_rent": {
            "type": "number",
            "title": "Monthly Rent"
          }
        },
        "type": "object",
        "required": [
          "block_id",
          "room_number",
          "floor",
          "capacity",
          "monthly_rent"
        ],
        "title": "HostelRoomCreate"
      },
      "HostelType": {
        "type": "string",
        "enum": [
          "BOYS",
          "GIRLS",
          "STAFF",
          "GUEST"
        ],
        "title": "HostelType"
      },
...
      "RoomAllocationRequest": {
        "properties": {
          "room_id": {
            "type": "integer",
            "title": "Room Id"
          },
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "bed_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bed Number"
          }
        },
        "type": "object",
        "required": [
          "room_id",
          "student_id"
        ],
        "title": "RoomAllocationRequest"
      },
      "RoomType": {
        "type": "string",
        "enum": [
          "CLASSROOM",
          "LAB",
          "SEMINAR_HALL",
          "AUDITORIUM",
          "LIBRARY",
          "STAFF_ROOM",
          "OFFICE",
          "OTHER",
          "NON_AC",
          "AC",
          "DELUXE"
        ],
        "title": "RoomType"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}